Remove useless intermediary filterBuilds
