+++ src/app/loadout-builder/LoadoutBuilder.m.scss
@@ -33,6 +33,10 @@
   :global(.search-filter) {
     min-width: 0;
   }
+
+  > * {
+    margin-top: 8px;
+  }
 }
 
 .processing {
