--- src/app/loadout-builder/filter/FilterBuilds.m.scss.d.ts
@@ -1,8 +1,0 @@
-// This file is automatically generated.
-// Please do not change this file!
-interface CssExports {
-  'filters': string;
-  'row': string;
-}
-export const cssExports: CssExports;
-export default cssExports;
