+++ src/app/loadout-builder/filter/LockArmorAndPerks.m.scss
@@ -12,7 +12,6 @@
 
 .area {
   padding: 4px;
-  margin-bottom: 8px;
   background-color: rgba(0, 0, 0, 0.4);
 
   &:global(.on-drag-enter),
