+++ src/app/loadout-builder/filter/TierSelect.m.scss
@@ -1,3 +1,5 @@
+@import '../../variables';
+
 .iconStat {
   display: inline-block;
   height: 17px;
@@ -18,4 +20,44 @@
 
 .hiddenOption {
   display: none;
+}
+
+.row {
+  display: grid;
+  grid-template-columns: min-content 1fr 45px 45px;
+  gap: 4px 4px;
+  margin-bottom: 2px;
+  padding-right: 4px;
+  background-color: black;
+  align-items: center;
+
+  > * {
+    margin: 4px 0;
+    @include phone-portrait {
+      margin: 8px 0;
+    }
+    &:first-child {
+      margin-left: 8px;
+    }
+    &:last-child {
+      margin-right: 4px;
+    }
+  }
+
+  label > span {
+    display: flex;
+    flex-direction: row;
+    align-items: center;
+    white-space: nowrap;
+  }
+
+  > select {
+    height: 18px;
+    padding-top: 0;
+    padding-bottom: 0;
+    padding-left: 4px;
+    padding-right: 4px;
+    background-position-x: calc(100% - 4px);
+    width: 45px;
+  }
 }
